requirements.txt
python-binance==1.0.19
pandas==2.1.0
numpy==1.24.3
requests==2.31.0
ta==0.10.2
python-dotenv==1.0.0
websocket-client==1.6.1